abstracted initialization of db (and server) from general express configuration
