+++ server/server.js
@@ -1,28 +1,9 @@
-require("dotenv").config();
-
-const express = require("express");
-const app = express();
-const mongoose = require("mongoose");
+const app = require("./app");
+const db = require("./db");
 const port = 5001;
 
-// Set up database connection
-mongoose.connect(process.env.DATABASE_URL, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true
+db.connect().then(() => {
+  app.listen(port, () => {
+    console.log(`Listening on port ${port}!`);
+  });
 });
-const db = mongoose.connection;
-db.on("error", error => console.error(error));
-db.once("open", () => console.log("connected to database"));
-
-app.use(express.json());
-
-const postsRouter = require("./src/routes/posts");
-app.use("/posts", postsRouter);
-
-app.get("/", (req, res) => res.send("Hello World!"));
-
-const server = app.listen(port, () =>
-  console.log(`Listening on port ${port}!`)
-);
-
-module.exports = server;
